v0.23.2 - Ajout de ChildPresence (21/01/2019) - Correction du deletepar array of id de StaffPresence (21/01/2019) - Ajout de tests pour StaffPresence (21/01/2019)

+++ src/Service/StaffPresenceService.php
@@ -106,8 +106,10 @@ class StaffPresenceService implements StaffPresenceServiceInterface
                 $object = $this->em->getRepository('App:StaffPresence')->findByData($staffPresence);
 
                 //Submits data
-                $this->mainService->delete($object);
-                $this->mainService->persist($object);
+                if ($object instanceof StaffPresence) {
+                    $this->mainService->delete($object);
+                    $this->mainService->persist($object);
+                }
             }
 
             return array(
